Record use/def link between names.
The DefTable type included a refTable field to record links between use
and definition Names, but it was unused. Now, whenever analysis code
calls lookupTypeDef or lookupObject, the link between use and definition
is store in the refTable. Type checking should fully populate this
table, because it needs to lookup all uses to ensure that their use is
well-typed.